requirements.txt
# LLM & Orchestration
langchain>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Speech to Text
openai-whisper

# YouTube
yt-dlp
youtube-transcript-api

# Vector Store


# Embeddings
sentence-transformers

# Web Search
duckduckgo-search

# Scraping
beautifulsoup4
requests

# A2A / API Servers
fastapi
uvicorn[standard]
httpx

# UI
streamlit

# Utilities
pydantic
